+++ lib/Db/CirclesMapper.php
@@ -360,7 +360,9 @@ class CirclesMapper extends Mapper {
 		$qb->insert(self::TABLENAME)
 		   ->setValue('name', $qb->createNamedParameter($circle->getName()))
 		   ->setValue('description', $qb->createNamedParameter($circle->getDescription()))
-		   ->setValue('type', $qb->createNamedParameter($circle->getType()));
+		   ->setValue('type', $qb->createNamedParameter($circle->getType()))
+		   ->setValue('creation', $qb->createFunction('NOW()'));
+
 		$qb->execute();
 		$circleId = $qb->getLastInsertId();