+++ lib/Db/MembersMapper.php
@@ -237,7 +237,9 @@ class MembersMapper extends Mapper {
 			   ->setValue('user_id', $qb->createNamedParameter($member->getUserId()))
 			   ->setValue('level', $qb->createNamedParameter($member->getLevel()))
 			   ->setValue('status', $qb->createNamedParameter($member->getStatus()))
-			   ->setValue('note', $qb->createNamedParameter($member->getNote()));
+			   ->setValue('note', $qb->createNamedParameter($member->getNote()))
+			   ->setValue('joined', $qb->createFunction('NOW()'));
+
 			$qb->execute();
 		} catch (UniqueConstraintViolationException $e) {
 			throw new MemberAlreadyExistsException(